requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bcrypt==4.3.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cognee==0.4.1
coloredlogs==15.0.1
cryptography==46.0.3
deprecated==1.3.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
eval-type-backport==0.3.1
fastapi==0.124.0
fastapi-users==14.0.2
fastapi-users-db-sqlalchemy==7.0.0
fastembed==0.6.0
fastjsonschema==2.21.2
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
instructor==1.13.0
invoke==2.2.1
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-core==5.9.1
kuzu==0.11.3
lance-namespace==0.0.21
lance-namespace-urllib3-client==0.2.1
lancedb==0.25.3
limits==4.8.0
litellm==1.80.7
loguru==0.7.3
makefun==1.16.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mistralai==1.9.11
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nbformat==5.10.4
networkx==3.6
nodeenv==1.9.1
numpy==2.3.5
onnxruntime==1.22.1
openai==2.9.0
packaging==24.2
pillow==11.3.0
platformdirs==4.5.1
playwright==1.56.0
pre-commit==4.5.0
propcache==0.4.1
protego==0.5.0
protobuf==6.33.2
pwdlib==0.2.1
py-rust-stemmers==0.1.5
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pylance==0.36.0
pympler==1.1
pyparsing==3.2.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
rdflib==7.1.4
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
ty==0.0.1a32
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.0
uvicorn==0.38.0
virtualenv==20.35.4
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
